+++ MICE/Info.plist
@@ -6,6 +6,8 @@
 	<string>$(SUPABASE_KEY)</string>
 	<key>SUPABASE_URL</key>
 	<string>$(SUPABASE_URL)</string>
+	<key>NSLocationWhenInUseUsageDescription</key>
+	<string>위치 기반 추천 및 서비스를 위해 위치정보를 사용합니다.</string>
 	<key>UIApplicationSceneManifest</key>
 	<dict>
 		<key>UIApplicationSupportsMultipleScenes</key>
